Add files via upload
pdo.php which contains code to connect to the database is absent for obvious security reasons. Though this code is in a raw form with comments, it is left as so for learner intentions. Also, the code is ready to go live. The salted pass in login.php will also be changed. see www.ngonike.dev/apps/profile-registry for live version.